Implement wrapper script for transparent file encryption (not working)
The scripts/sealing script first unseals the files specified in
/etc/python-sgx/sealing_manifest, executes the command provided on
the command line, and then seals the files again. Also seals whole
directories and removes sealed files if the unsealed file gets deleted.

Unfortunately, Graphene currently doesn't support ephemeral filesystems,
so we have to use a chrooted directory in order to be able to store the
unsealed files. This of course also allows the OS access to the unsealed
files, rendering the whole thing useless.

Even if Graphene would support ephemeral filesystems at some point, this
script would only work if the child process, which is being executed in a
separate enclave, would share the ephemeral filesystem with the parent.
Else the parent unseals the files in its own filesystem and the child would
not be able to access it.

Surprisingly, there is already a similar issue at the moment, even though
Graphene does not use ephemeral filesystems. But somehow the modifications
in the child are not synced with the parent, so when the unmodified files
of the parent get sealed, the modifications of the child are lost.